+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Create API to get count of accounts matching criteria [#619](https://github.com/rokwire/core-building-block/issues/619)
 ### Fixed
 - Panic on anonymous refresh [#621](https://github.com/rokwire/core-building-block/issues/621)